refactor(web): colocate single-consumer types, delete dead type files
- Delete publish-types.ts: 4 type aliases had zero consumers, move
  isCredentialError() into deploy-store.ts (its sole consumer)
- Delete subagent-types.ts: move types into fetch-subagent-detail.ts
  (its sole consumer)
- Tighten CONVENTIONS.md types/ guidance: types live with their owning
  module, types/ is only for cross-domain types with no clear owner
