+++ apps/web/src/utils/fetch-subagent-detail.ts
@@ -2,14 +2,42 @@
  * Fetch subagent execution detail from the daemon.
  *
  * The daemon route schema declares `events: z.array(z.unknown())` so the
- * generated SDK types events as `Array<unknown>`. This wrapper casts to
- * the known runtime shape (`SubagentDetailResponse`).
+ * generated SDK types events as `Array<unknown>`. This module defines the
+ * known runtime shape and casts to it.
  */
 
 import * as Sentry from "@sentry/browser";
 
 import { subagentsByIdGet } from "@/generated/daemon/sdk.gen";
-import type { SubagentDetailResponse } from "@/types/subagent-types";
+
+// ---------------------------------------------------------------------------
+// Types — mirrors `parseSubagentMessages()` in the daemon's
+// subagents-routes.ts. Kept here because this fetch wrapper is the sole
+// consumer and the SDK's `z.unknown()` erases the real shape.
+// ---------------------------------------------------------------------------
+
+export interface SubagentEvent {
+  type: string;
+  content: string;
+  toolName?: string;
+  isError?: boolean;
+  messageId?: string;
+  text?: string;
+  result?: string;
+  timestamp?: number;
+}
+
+export interface SubagentDetailResponse {
+  subagentId: string;
+  objective: string;
+  status?: string;
+  usage?: {
+    inputTokens: number;
+    outputTokens: number;
+    estimatedCost: number;
+  };
+  events: SubagentEvent[];
+}
 
 export async function fetchSubagentDetail(
   assistantId: string,
